todo_app/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
future==0.18.2
pytz==2022.6
sqlparse==0.4.3
